Changes:
There are 100 doors in a row and all closed initially.
You make 100 passes by the doors starting with the first door every time.
The first time you visit every door and toggle it (if the door is closed, you open it, if it's open, you close it).
The second time you visit every 2nd door only (door #2, #4, #6). 
The third time, every 3rd door (door #3, #6, #9), and so on until you visited 100th time.
What state are the doors in after the last(100th) pass? Which are open which are closed?

Source: Internet
Solution: https://www.geeksforgeeks.org/puzzle-16-100-doors/